refactored arc2 to its own file
